REFAC 003: Editted U-Test 012
The test made a copy of the array of cards, referencing the same array. Thus after shuffling, both arrays in the test pointed to the same array and were never different. The arrays are now defined seperately: Both start as the default sorted array, and one is shuffled. Then they are compared.
Test fails