CODE 013: All cards are correctly added to each players hand.
Cards now have a 'taken' value that indicate that they're within a players hand. This value is checked for, so that cards do not appear within more than one players hand.

While playercount is capped at 5, the maximum amount of cards in play is 60. Thus we don't have to worry about running out of cards, so long as they are correctly taken out of players hands.